backend/requirements.txt
Django==6.0.4
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
python-dotenv==1.2.2
pandas==3.0.2
numpy==2.4.4
joblib==1.5.3
shap==0.51.0
xgboost==3.2.0
psycopg2-binary==2.9.12

# Deployment/runtime
gunicorn>=23.0.0

# Optional for GCP bucket-backed media storage
django-storages>=1.14.0
google-cloud-storage>=2.18.0
